+++ lib/src/time/calendar/holidays/juneteenth.dart
@@ -3,12 +3,13 @@ import 'package:timezone/timezone.dart';
 import '../holiday.dart';
 
 /// Independence day is a federal holiday.
-class Juneteenth implements Holiday {
-  @override
-  HolidayType holidayType = HolidayType.juneteenth;
+class Juneteenth extends Holiday {
+
+  Juneteenth() {
+    holidayType = HolidayType.juneteenth;
+  }
 
   @override
-
   /// Before 2021-06-19, this function will return null as the holiday
   /// didn't exist
   Date? forYear(int year, {required Location location}) {