+++ lib/src/time/calendar/holidays/patriots_day.dart
@@ -3,11 +3,12 @@ import 'package:date/date.dart';
 import '../holiday.dart';
 import '../holiday_utils.dart';
 
-/// Patriots' Day. See https://en.wikipedia.org/wiki/Election_Day_(United_States)
-/// Set on 3rd Monday in April.  It's a state holiday in MA and ME.
-class PatriotsDay implements Holiday {
-  @override
-  HolidayType holidayType = HolidayType.patriotsDay;
+class PatriotsDay extends Holiday {
+  /// Patriots' Day. See https://en.wikipedia.org/wiki/Election_Day_(United_States)
+  /// Set on 3rd Monday in April.  It's a state holiday in MA and ME.
+  PatriotsDay() {
+    holidayType = HolidayType.patriotsDay;
+  }
 
   @override
   Date forYear(int year, {required Location location}) {